Warmup caches before counting cycles
This seems to remove the noise in some of the vlseg/vlse results, and I'm able to get two pretty consistent consecutive runs now for vlse_LMUL_x_SEW_throughput.

My guess is that because the measure function is so big it's mostly the instruction cache that was causing the variance beforehand.